rename: 폴더명 컨벤션 통일
