+++ app/courses/new/page.tsx
@@ -1,4 +1,4 @@
-import AddCoursePlan from '@/src/widgets/AddCoursePlan'
+import AddCoursePlan from '@/src/widgets/add-course-plan'
 
 export default function CourseNew() {
   return <AddCoursePlan type='course' />
