+++ app/layout.tsx
@@ -1,8 +1,8 @@
 import type { Metadata } from 'next'
 import './globals.css'
 import Spacer from '@/src/shared/ui/Spacer'
-import DefaultHeader from '@/src/widgets/DefaultHeader'
-import DefaultFooter from '@/src/widgets/DefaultFooter'
+import DefaultHeader from '@/src/widgets/default-header'
+import DefaultFooter from '@/src/widgets/default-footer'
 import localFont from 'next/font/local'
 import { AnimatePresence } from 'framer-motion'
 import { ConfigProvider } from 'antd'
