+++ src/views/detail-user/index.tsx
@@ -7,8 +7,8 @@ import ListUserPlace from '@/src/features/plan/list-user-place'
 import ListUserCourse from '@/src/features/course/list-user-course'
 import { Select } from 'antd'
 import Spacer from '@/src/shared/ui/Spacer'
-import FloatingWriteButton from '@/src/widgets/FloatingWriteButton'
-import Header from '@/src/widgets/Header'
+import FloatingWriteButton from '@/src/widgets/floating-write-btn'
+import Header from '@/src/widgets/header'
 import type { CourseType } from '@/src/entities/course/type'
 import type { PlaceType } from '@/src/entities/place/type'
 import type { UserType } from '@/src/entities/user/type'
