+++ src/views/update-plan/index.tsx
@@ -6,7 +6,7 @@ import KakaoMap from '@/src/shared/ui/KakaoMap'
 import { DatePicker } from 'antd'
 import SearchPlace from '@/src/views/search-place'
 import { useState } from 'react'
-import DragPlace from '@/src/widgets/DragPlace'
+import DragPlace from '@/src/widgets/drag-place'
 import dayjs from 'dayjs'
 import customParseFormat from 'dayjs/plugin/customParseFormat'
 import { CourseType } from '@/src/entities/course/type'
