+++ src/widgets/course-plan-layout/index.tsx
