+++ src/widgets/default-footer/index.tsx
