+++ src/widgets/drag-place/index.tsx
