+++ src/widgets/floating-write-btn/index.tsx
