+++ tsconfig.json
@@ -34,8 +34,8 @@
     "next.config.js",
     "src/views/detail-comment",
     "src/shared/ui/ProfileImage",
-    "src/widgets/CoursePlanLayout",
-    "src/widgets/FloatingWriteButton"
+    "src/widgets/course-plan-layout",
+    "src/widgets/floating-write-btn"
   ],
   "exclude": ["node_modules"]
 }
